refactor: 비즈니스 로직과 서비스 로직 분리
