+++ src/shared/ui/FileUploader/index.tsx
@@ -1,21 +1,16 @@
 'use client';
 
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 
 import { FileType } from '@/entities/file/model/file';
 import Chain from '@/shared/asset/svg/Chain';
+import useFileUploaderState from '@/shared/model/useHandlingFileUploader';
 import FileList from '@/shared/ui/FileList';
 
 interface FileUploaderProps extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'value'> {
   label: string;
   uploadedFiles?: FileType | FileType[];
   isMultiple?: boolean;
-}
-
-interface LocalFile {
-  id: string;
-  name: string;
-  file: File;
 }
 
 interface NewFileInputProps {
@@ -50,81 +45,20 @@ export default function FileUploader({
   isMultiple = false,
   ...props
 }: FileUploaderProps) {
-  const [existingFiles, setExistingFiles] = useState<FileType[]>([]);
-  const [newFiles, setNewFiles] = useState<LocalFile[]>([]);
+  const {
+    existingFiles,
+    newFiles,
+    displayFiles,
+    buttonText,
+    handleFileChange,
+    handleRemoveFile,
+  } = useFileUploaderState({ uploadedFiles, isMultiple });
 
   const inputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    if (!uploadedFiles) {
-      setExistingFiles([]);
-      return;
-    }
-    const normalizedFiles = Array.isArray(uploadedFiles) ? uploadedFiles : [uploadedFiles];
-    setExistingFiles(normalizedFiles);
-  }, [uploadedFiles]);
 
   const openFileDialog = () => {
     inputRef.current?.click();
   };
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedFiles = e.target.files;
-    if (!selectedFiles || selectedFiles.length === 0) return;
-
-    if (!isMultiple) {
-      const file = selectedFiles[0];
-      setExistingFiles([]);
-      setNewFiles([
-        {
-          id: crypto.randomUUID(),
-          name: file.name,
-          file: file,
-        },
-      ]);
-    } else {
-      const fileArray = Array.from(selectedFiles);
-      const localFiles: LocalFile[] = fileArray.map((file) => ({
-        id: crypto.randomUUID(),
-        name: file.name,
-        file: file,
-      }));
-      setNewFiles((prev) => [...prev, ...localFiles]);
-    }
-
-    e.target.value = '';
-  };
-
-  const handleRemoveFile = (id: number | string) => {
-    if (typeof id === 'number') {
-      setExistingFiles((prev) => prev.filter((file) => file.id !== id));
-    } else {
-      setNewFiles((prev) => prev.filter((file) => file.id !== id));
-    }
-  };
-
-  const getButtonText = () => {
-    const totalCount = existingFiles.length + newFiles.length;
-    if (totalCount === 0) return '파일 첨부';
-
-    if (!isMultiple) {
-      if (newFiles.length > 0) return newFiles[0].name;
-      if (existingFiles.length > 0) return existingFiles[0].originalName;
-      return '파일 선택됨';
-    }
-
-    return `${totalCount}개의 파일 첨부됨`;
-  };
-
-  const displayFiles: FileType[] = [
-    ...existingFiles,
-    ...newFiles.map((f) => ({
-      id: f.id,
-      originalName: f.name,
-      storeName: '',
-      uri: '',
-    })),
-  ];
 
   return (
     <div>
@@ -142,9 +76,9 @@ export default function FileUploader({
           </span>
           <span
             className="max-w-[220px] truncate text-sm text-gray-400"
-            title={getButtonText()}
+            title={buttonText}
           >
-            {getButtonText()}
+            {buttonText}
           </span>
 
           <input
